+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file is used to list changes made in each version of cookbook-openstack-net
 * Allow dhcp_delete_namespaces and router_dhcp_namespaces to be overridden.
 * Add support for openvswitch agent MTU size of veth interfaces
 * fix fauxhai version for suse and redhat
+* Allow rootwrap.conf attributes
 
 ## 9.1.0
 * python_packages database client attributes have been migrated to